I/O: exposure-key change for io backends
The functional change is that we only over-write the EXPOSURE keyword if
the attribute is not None. This means that any value set in the header
will be used, and ensures that we don't accidentally set the keyword to
"None", which is not a universally-accepted value for "no data" in FITS
parsers.

There are a few minor code cleanups, mainly to be explicit about the
header keyword (should be no functional changes from these).

There are some minor documentation changes, moving references in-line
rather than as an actual reference block.