I/O: add a check of write_pha, write_rmf, and write_arf
The idea is that we can load in a PHA file plus responses
and do a simple fit and get the same answer as after we
write out the PHA, background PHA, RMF, and ARF and then
use that for the fit. This shouldn't be needed if the individual
write routines have been well tested, but it's a nice functional
check.

The actual fit isn't that important, as long as it's repeatable.