I/O: strip FITS headers of structural keys on read
We have long had subtle differences in the headers of objects
read in by crates and AstroPy. This is always going to happen,
but we can reduce this by stripping out "structural" keys from
the FITS header. Note that crates already does this for some
keywords, but this is more extensive (but is likely to need
some tweaking over time).

What this means is that we lose some information - e.g. if a
WCS transform has been added to a column, as we see in the XMM
RGS RMF case, but it is needed to make it easier to write out
FITS files that have previously been read in. An alternative is
to be much more careful about all the keywors that apply to the
columns (or image) being written out, but we still have to worry
about keywords referencing non-existant columns, and replicating
this logic everywhere is is needed. It is just easier to remove
the values here when the file is read in.

This potentially means that a user can add them manually after the
file has been read in, and so they may get written out, but that
depends on the I/O backend (i.e. is unlikely to work for crates).

If users need access to this metadata then they will have to
read the data in using their preferred I/O library and access it
separately.

Note that we already did some of this when reading in image files,
to strip out the 2D WCS transforms. This is now split between the
new code and some repeated code in the two backends (since it's
not clear how we want to handle the "secondary" WCS transforms
like the "P" ones that CIAO uses).