XSPEC: fix #1814
In #1259 (Sep 2021) I updated the XSPEC models to generally use
the XSParameter rather than Parameter class for parameters. This
was to allow users to over-ride the hard min/max ranges, as this
is a feature in XSPEC used by a number of models which use a
value < 0 (say, where 0 is the minimum parameter value) to change
behavior in some way.

However, some parameters were marked as "do not allow the limit
to change" (more like the existing Parameter hard limits) using
the XSBaseParameter class. This is used for table models where
we had experience with crashes and memory corruption if the
parameter range were exceeded.

One of the parameters that can - but not always - get added to
a XSPEC table model is the redshift parameter, which currently
XSPEC decided to set to 0 with a range of 0 to 5. This was set
as a XSBaseParameter to match the other table parameters, but
there's less need to do this, as it is not-as-directly related
to memory-crashing bugginess as the other parameters. And a user
was setting values to outsize the z=0 to 5 range and the 4.15
release where all this was added broke things.

So we switch just the redshift parameter from XSBaseParameter
to XSParameter. Unfortunately we do not have a XSPEC table
model we can use to check this out with (in the tests at least,
there are models such as the additive table model

https://sites.astro.caltech.edu/~mislavb/download/borus02_v170323/borus02_v170323a.fits

that we can use to test).