XSPEC: support ESCALE keyword in XSPEC table models
Since we can not guarantee that ESCALE is present, unlike the other
keywords, we have to rework the keyword code slightly. More could
be done, but that would lead to changing the I/O interface which
I do not want to do here.

We also add an explict check to error out if NXFLTEXP is set (and
more than 1). It is unclear what should be done if NXFLTEXP = 1.
We can not support these models until we support the XFLT database.