XSPEC: update model parameters to match XSPEC 12.13.1 / HEASOFT 6.32
These updates were pointed out by running scripts/check_xspec_update.py
on the model.dat from HEASOFT 6.32 (XSPEC version 12.13.1). I am not
sure when these changes were first made to XSPEC, but they are mainly
"documentation" style changes so I do not worry about using them in
a build with an XSPEC library before these changes were made.

The two changes are

- reducing the maximum for the Valocity parameter from 1e6 to 1e4 km/s
- updating the ismabs model to use the latest version of the XSPEC
  parameter names

For the latter, I noted to HEASARC quite some time ago that there
existing naming scheme (of <elemement><species>) lead to a confusion
for the Si and S parameters for a system like Sherpa where the
parameter names are case insensitive, since is "sii" meant to be
"SiI" or "SII". So we called these two parameters "Si_species" and
"S_species". HEASARC have finally applied this to all the elements
(apart from H and Fe), so we make this change as well and use the
"aliases" capability of the model to allow the old names to be used.

I had added a test to check the old behavior, so this test had to
be updated to follow the new behavior, and I added in a check that
the deprecation warning is created when an "old" parameter name is
used.

Note that the parameter name is ignored by the XSPEC model library,
as all the code cares about is getting an array of npar numbers, so
we can actually call these parameters whatever we want, as long as
the ordering matches.

I added a note for the bwcycl model to point out that we are not
going to follow the XSPEC naming of the del parameter because we
can not. My guess is that this parameter has always been called
del and we decided to use delta, rather than it changing from
delta to del at some point.