PHA: allow tabStops to be a list, not just a ndarray
Since the C++ code assumes the tabStops argument is a ndarray,
force it to be one. An alternative approach is to fix up the
C++ code to do the conversion but that is significantly more
work (and one can view the Python code as being tasked with
sending in the correct data, so it isnot a terrible place to do
the conversion).