PHA: group_xxx defaults to using the existing filter
This is a breaking change. To get the pre 4.16.0 behavior add a
tabStops argument to the group_xxx set to an array of zeros with
a length that matches the number of channels. So, for ACIS imaging
mode spectra, something like

  [ui]    group_counts(20, tabStops=[0] * 1024)
  [obj]   pha.group_counts(20, tabStops=np.zeros(1024))

The change here is that when tabStops is None - which is the
default behavior - the code will now use the inverse of the
existing mask as the tabStops argument. This means that the
grouping will only be done on the filtered data, and the data
outside the filter will be ungrouped (but also unused until a
notice call is made to extend the range).

This should help filter ranges "growing" by accident, since
a set of calls like

    load_pha("src.pi")
    notice(0.5, 6)
    group_counts(15)
    group_counts(20)

will now lead to no change in the noticed range, whereas before
it would likely increase at both the low and high edges (depending
on where the grouped edges change).

Fix #956