Tests: add tests related to #1830
The underlying issue for #1830 is that the pyfits backend can not
read in the HRC-I RMF correctly, so we can have an explicit check
of this in sherpa/astro/io/tests/test_io_pha.py. We also add a
more "functional" test where we check we can load in a PHA file
and evaluate a model, and compare it to data - this is in
sherpa/astro/ui/stata/test_astro_ui_io.py.

I have checked that these tests pass with a crates backend and are
marked as XFAIL with the pyfits backend.