Tests: check XSPEC table model with ESCALE and REDSHIFT
In #1857 I noted that XSPEC 12.13.1 and earlier created table
models that failed when both REDSHIFT and ESCALE keywords were
set (because they create a z parameter and then an escale one).
I reported this problem and it should be fixed in XSPEC 12.13.1a
and is an easy fix to apply, but let's make sure we have a test
before we make the fix.

I also took the advantage of reworking some related tests to make
it obvious when the table models should be returning the same
data.