plot: remove sherpa.plots.utils.intersperse
This routine was the cause of issue #1828 - which was, to be
honest, a corner cause - and is no-longer used. It could be
marked as deprecated, but this is a relatively new module and
also very low-level, so we feel it is okay to remove it without
a deprecation period (particularly as this is not a hard
algorithm to implement or replace).