I/O: allow XSPEC table models to be written out
Allow XSPEC table models to be written out. The interface is
labelled as experimental as I am not 100% convinced it is
useful (in particular, there are different ways we can send
in the paramter values used to set up the model spectra).

I have also added these routines to a new module - that is
sherpa.astro.io.xstable - rather than to sherpa.astro.io
because

a) it's still experimental
b) even when it's not experimental, it is still a highly
   specialized piece of code that I do not think we need
   provied by sherpa.astro.io or sherpa.astro.ui[.utils]

This introduces a number of classes that might be useful for
other data file types, but for now we leave them in
sherpa/astro/io/xstable.py. It does add a "generic" set_hdus
routine to the backend, which only supports FITS output and - at
present - is limited to table-only data. Other routines could be
written to use set_hdus rather than the type-specific version
(or, make the type-specific one be a simple call to set_hdus
when FITS is selected), but this is also left for another time
(as it requires moving the types from xstable.py).