IO: minor clean up of the backends
This is mainly a cleanup, but there is a small change to how
the pyfits backend works: when checking for ARF files and the
block name is not SPECRESP then we now check for both
HDUCLAS1 and HDUCLAS2 being the expected values, that is
"RESPONSE" and "SPECRESP" rather than checking one of them
matches "RESPONSE". This means that we now error out earlier
if we try to load a RMF as an ARF (which should have "RESPONSE"
and "RSP_MATRIX"), as seen in one of the tests for this
situation.

There is also some cleanup of the crates backend to support
the current capabilities of crates and remove some un-needed
code.