np: replace string_ use
In most cases we just use np.bytes_ (needed for NumPy 2.0) along
with a note to check whether we still need this (historically we
have had issues from crates returning bytes rather than strings,
but it's not clear if this is still present).

One case I just removed it as it was only for a warning message and
it's not clear why it was being used.

There are several minor code cleanups related to this: use of
isinstance() rather than "type(..) in ...".

A few minor drive-by code clean ups to change

  if blah:
    do_something
  else:
    raise an_error

to

  if not blah:
    raise an_error

  do_something