Tests: add an example of issue #1873
The problem is that we do not include the "set_psf" call in the
ASCII serialization. Since it's not obvious what the text will be
we can't check the output serialization, but we can at least
check that the PSF has not been applied when we restore the
saved file.