Tests: improve coverage of save_all
Check
  - linked parameters
  - PHA with no response
  - PHA with type=counts (as part of the previous test)
  - PHA from CSC
  - "full model" handling (currently XFAIL)
  - table model
  - xspec table model (currently a regression test as the
    code does not create a valid save file)
  - pileup model (currently XFAIL because of inability to
    recreate data not read from a file)
  - dataspace1d/2d use (currently XFAIL)
  - example of load_data with an ascii file, which has lead
    to viwing interesting differences in the two I/O backends
  - PHA grating dataset (multiple backgrounds) using wavelength
    XFAIL see #320
  - PHA2 file XFAIL see #1882 (issue #320 plus more)

For those tests where we know they fail because the save_all
output is wrong in some way, we ensure they fail in the
expected manner. This is to make sure that they don't suddenly
start failing for a different reason (due to an error elsewhere)
which we don't see. This will only be noted when running with
the --runxfail pytest flag.