ui: initial support for load_arrays in save_all
Add basic support for calling save_all for a session where one
or more datasets were created manually - e.g. with load_arrays,
dataspace1d, dataspace2d, or set_data.

This only provides basic support, as things like header keywords
are not copied over and there are likely PHA settings that are
not handled.

The aim is to provide basic support and hope that is enough
to be usable, with additions added later. This really needs
changes to the Data object class, to identify when a dataset
has been changed from it's original / read-from-file values,
as

a) we can simplify a lot of the pha code if we don't need to
   reset grouping/quality

b) there are cases where a used can have tweaked a file
   and it would be nice to catch that

although I think there will always be possible problems (e.g.
an on-disk change to the file), so this will never be perfect.

The DataIMG restoration is more-complicated since we have to
worry about data where the coord has been changed from logical,
since we need to send in logical coordinates to the DataIMG
constructor. Fortunately this is relatively easy (although
testing shows some unrelated issues like #1880 and #1789).