Tests: better handle case where both astropy and pyfits are available
It used to be hard to check what backend was in use, for those
tests that need to behave differently with a particular backend
(or, rather, I hadn't realised we could use the backend name as
a way to decide), but with @hamogu's backend changes it is a lot
easier.

So this commit just cleans up a test case that fails if you
happen to have crates and astropy installed and the I/O backend
is set to pyfits, because the test was essentially wrong. We
now check for the backend in use rather than jsut checking what
packages are installed/available.

This is only really an issue for people building against
CIAO.