ui: change import ordering for save_all
We now make sure that all imports are done first. This is only
relevant for cases which need to write out manually-created image
files, and so need to also import the sherpa.astro.io.wcs.WCS
symbol.

It is not clear whether we require this symbol to be always
present or it is optional (this is only relevant for the
code generation, not the code created by save_all).

This is actually a large code change, as we now create the output
contents before writing it out to a file. There's more that
could be done [*], but for now leave as just this change.

[*] for instance, the "output state" could be built up by a
    class rather than just manually adding to a dictionary,
    but we don't yet need the class (e.g. to over-ride it)