ui: save_all limits model output when not needed
This avoids writing out banners for components and model
expressions, and removes some blank lines that were being
added.

The approach taken to avoiding printing out the banned is
rather hacky, but it works. The code could be re-written
to only add the banner if we know we need it, but that is
a larger change I didn't want to make here.