ui: save_all internal clean up
Separate out some of the code for setting up datasets, and take
advantage of knowing we have a DataPHA object to avoid quite so
many try/except blocks. We know prefer to send the DataPHA
object around, and use that API rather than the UI API, to
access the data we need.

As part of this we have removed more "blank" sections for PHA
files (no semantic changes).