fix up an error from a rebase due to a conflict
When I rebased on main, I left in a "requires_plotting" that has been removed in this branch.
However, there are some many recent changes to those
import statements, that I can't find now which commit is the one I need ot fix up,
so I'll just  put this in a separate commit.