Tests: change some XFAIL tests to regression tests
In changing save_all I have found it useful to check that we are
still failing in the same way for issues #320 and #1882, so make
the current tests regression tests (so they pass, even though the
output is incorrect) rather than mark them as xfail.

We do not do this for all XFAIL tests.