Unify set_backend and TemporaryPlottingBackend
Both ways of setting the backend (permanently or temporary) now accept
the same arguments, issue the same errors, and are tested.

A few unrelated fixes to docstrings are included in this commit in order to make
doctestplus pass and thus allow testing of the examples in the docstring for the two functions.