UI: set_analysis reports the filter change - fix #1590
This follows notice/ignore (#1562) and group_xxx (#1637) changes
to report the filter for any dataset that set_analysis operates
on. Unfortunately it can't easily use the code used in the previous
PRs since that doesn't support the "analysis units" changing,
which is the main reason for set_analysis! So the code does it
manually, but it's not a lot of code.