instrument: convert a RMF to an image (set of pixels)
Adds rmg_to_matrix (very low level) and rmf_to_image (which
returns a DataIMG to fit into the sherpa data hierarchy).
The image can be written out using the standard sherpa.astro.io
routines. These routines have not been added to the UI layer,
that is, there's no wrappers around this, but it's not obvious
what the UX should be, so we can think about adding this in a
later PR (or never do; there's an argument to be had to say
it should stay in the instrument module).

As part of this I introduce the RMFMatrix dataclass as a way of
sending around "structured" data rather than just deal with a tuple
of values. This has the advantage of being easily able to add
validation - e.g. check the sizes match - and is part of
issue #1893. I do note though that this is a "very limited use"
item - it's used in rmf_to_image and is exposed as we make
rmf_to_matrix a part of the API - but I think it's worth the
effort.

In part because of the use of the dataclass (the need to annotate
the fields) this adds typing of the new routines, along with
updates to the code to let 'mypy sherpa/astro/instrument.py' passes,
but most of the code remains un-typed.