Change default string representation to assume numpy
Note that this works for numpy functions, but not for non-numpy functions.
They will still be printed in the string representions, but not with sufficient information
to re-create the function from e.g. a save_all file. This is hard to solve in general and we expect it to be a very rare use case.