Tests: improve the read-a-rmf-image test and show issue #1889
The test has been changed to be a test of the lower-level code,
as warrented by the test file, rather than the Session code.
The Session layer is a very-thin layer above the code tested
here.

Ideally we'd have kept the RMF that was used with rmfimg to
create test_rmfimg.fits, but we can get close to it using CSC 1.1
(the revision number of the file is 3 rather than 2 but the
pixel values generated by rmfimg match which suggests it was
likely an unrelated metadata change for version 3). Using this
RMF I have manually convolved a constant model (with c0 set to
100) and got several values we can use to test the RMF we
create from test_rmfimg.fits. Of course the existing code, because
it did not have such a round-trip case, gets the number of channels
wrong (assuming 900 rather than 1024) which is issue #1889.