instrument: ensure RMF need not be square fix #1889
Add a small amount of data validation to DataRMF to show some
egregious problems (more could be done, and a decision needs
to be made if fields can only be set when calling __init__,
but those are larger-scale changes that need some thinking
about, so this is a minimal change).

The sherpa.astro.instruent.create_non_delta_rmf routine now
no-longer assumes that the number of channels matches the
RMF grid. This should fix #1889.