ui: add calc_bkg_stat - fix #333
Adds the routines

  calc_bkg_stat
  calc_bkg_stat_info
  get_bkg_stat_info

to the astro UI session. This is a relatively-easy extension of
the existing cacl/get_stat[_info] routines. There is one small
issue, of what to do when you have multiple background datasets
with different background ids, but for now just list all such
ids and let the user worry about the possible different mappings.