data: group backgrounds before source - fix #1881
With the way the groupnig was working, we tried to group the
backgrounds before they necessarily had any grouping flags. This
commit changes the order.

The 4.15.1 release added #1660 which was to fix #1657, which made
sure that the background data was grouped whenever the source data
was grouped. So, this update is only due to a recent behavior change,
and not something more serious (which was my original worry).

The group_xxx commands all have the same structure:

    loop over backgrounds
        call the method on the background

    call _dynamic_group to apply to the data

and I thought about abstracting that (probably into _dynamic_group),
but I decided against it as the code repitition isn't terribly large
and although the high-level description is easy to abstract, the
actual code would be harder to do so (reducing the benefit of the
abstraction).