Tests: add a test for #1881
This is basically the test case discussed in issue #1881 but
done with DataPHA objects rather than reading in from a file.

Technically we should test all the group_xxx routines, but
as the logic is the same for them all we just test one of them.
I chose group_counts since it's likely to be the most-used and
because it creates different grouping for source and background,
which is useful to check for.