UI: ensure rng parameter for method parameters is set (if set)
We have an issue with the UI layer in that the RNG set in
set_rng is not getting passed down to the optimisation method.

There are a number of possible approaches, but none particularly
obvious, so go with the simplest possible approach:

- whenever the method is set, update the rng option to match
  the existing set_rng value **if** the rng option exists

This allows a user to over-ride the setting, but it requires
that they explicitly set the rng option whenever the method
is changed to "moncar" - i.e.

    set_method("moncar")
    set_method_opt("rng", <whatever they want it to be>)

but it is assumed that most people would want the value to
track the set_rng call.

Or course, if you

    set_rng(rng1)
    set_method("moncar")
    set_rng(rng2)

then would you expect the moncar rng to get updated to? I
have decided that we do not have enough experience to make
a sensible choice here, so I do not try to be clever (e.g.
this does not make the set_rng call reset the optmethod rng
parameter).