opt: montecarlo clean up of seed/rng handling
With the introduction of the "rng" code the seed handling in
the o[timisation code is a bit confusing. It didn't help that I
added more confusion, so simplify the seed argument into

- set to None, then we use rng to generate a value
   - this always used to be the case, we just never documented it,
     but the code used to be

        random.seed(seed)
        if seed is None:
            # pow(2,31) == 2147483648L
            seed = random.randint(0, 2147483648)

- otherwise we use the seed vlaue

As noted in earlier commits for the "rng" approach, we need to
keep the seed argument for the low-level code, as it is passed
to C++ code. We **could** just not accept the rng parameter for
the montecarlo routine, we in that case the seed parameter would
have to be set (i.e. not able to be set to None). This is a
very-attractive thought, but we then would require the user of
the UI layer to create a seed value (if they didn't want to
use the default value) where they could always just set the "seed"
parameter to None.