crates: internal clean up
This module was written when crates was very young, and then
accreted fixes and updates to work around issues. However, it
hasn't really been re-worked to use the latest capabilities
of crates. This commit makes some changes, mainly

- consoloating functionality (mainly in the header handling)
- remove unused routines
- avoid "hasattr" type checks when we know that the given
  data structures contain the given attribute (presumably from
  initial testing of the "varlen" support)
- move to using methods on the object rather than top-level
  crates routines; this can be surprising as the two versions
  have different characteritics, in paticular in error handling
  (e.g. do we return None or raise an error)

There is more that can be done, but that would require much-more
significant restructuring that is not worth it as this time.

Some documentation has been addde, but more is needed.