Move types into sherpa.utils.numeric_types module
As suggested by @hamogu in the code review.

This is a lot of churn - including some unnescessary, but pleasing
to me, re-arrangement of the import order to try and be more
alphabetical - for no significant change. However, it does make it
more obvious where the type definitions on the Python side are.

I do winder if the Int/UInt types should be sized rather than
intp/uintp, but that's for a different PR.