Display legends in matplotlib by default if set
This change is useful in it's own right, but it is needed to make the RMF Plots for the rich display in the notebook
with our usual plotting backend infrastructure. Those changes are in the next commit, and will include a case where the legend is actually used in th tests.