ui: clean up documentation on internal include_bkg_data variable
The aim of this PR was to improve the documetantation and behavior
of the internal include_bkg_data flag, as it was created and then
re-created with two different comments. The second version as been
removed (and there are tests that check this code path).

In doing this I changed the routine to use "pha" rather than "d"
as the variable name - to avoid single-letter names and to pick
a variable name a bit-better matched to it's "meaning" (and also
"m" to "mdl").