Tests: add a test of a multi-matrix RMF
At the moment we

a) only use one of the matrices
   NOTE: it isn't always the first matrix, as it depends on
   how the backend behaves (and crates ends up chosing the
   second matrix, at least for the test file we create)

b) do not tell the user we are skipping the extra matrices

As we don't have a multi-matrix RMF for sherpa-test-data we
manually create one - using some of the new code for this added
to the I/O layer. This code doesn't let us easily set up the TLMIN
value for the RMF, so when we read it back in we expect to see a
warning about it.