IO: unify backend support of TLMIN in strange RMF case
The existing code tries to take the RMF offset from the CHANNEL
column of the EBOUNDS block when it's not been set via the
F_CHAN column of the MATRIX block. This commit attempts to
rationalize the behaviour of the two backends.

We should probably just remove this completely, but there's
plenty of other stuff to worry about first (such as what to
do with files with missing data), so we just make sure we have
equality (ish) in behaviour.