IO: add RMF types to io_types
The RMFMatrixData and RMFEboundsData types are used to pass
information from the backend to the IO layer. The aims are

- to better document the fields
- to support multiple matrices, even though that is not yet
  handled

There is some change in behaviour over partially-filled RMF
files: e.g. ig EBOUNDS is set we know require E_MIN, E_MAX, and
CHANNEL columns whereas before we could provide only some of
the data. As mentioned in #1871 this support doesn't make much
sense and just complicates the backend code.

One slightly-more significant change is that this centralizes
the handling of the "TLMIN4" keyword value for RMF files - that
is, does the file correctly tell us what the "starting value" is
for channels. The OGIP spec says this is the TLMIN for the F_CHAN
column in the MATRIX block, which is normally TLMIN4 for the FITS
officianado. However, we have one XMM RMF which provided this
information using the CHANNEL column of the EBOUNDS channel (it
is a sensible place for it, it is just not in the OGIP spec!). So
we now

- check both (we wre already doing this)
- only warn if they are bith missing
  - this is new, and means that users may see slightly-less
    warning messages from reading in the response