IO: allow multi-matrix RMF files to be read in and used
So, we treat a multi-matrix RMF as a special RMF, rather than
dealing with it at the instrument layer (i.e. not making use of
MultipleResponse1D/MultiResponseSumModel) because the chosen
approach makes the smallest amount of changes to the normal
UX - that is, a user can "load_rmf" a single file, or set the
RESPFILE keyword to it and it all should just work.

We may want to change this, but that is a longer-term project
(e.g. see #1907 and related issues), and it will come down to
experience with using multi-matrix blocks and other changes we
want to make to better handle multiple responses (e.g. for
grating data).

Note that the new DataMatrixRMF class

- does not include the special-cased filtering behaviour
  seen with DataRMF. It could be done, but I just wanted to
  get the basics done first

- does not "display" nicely - e.g. the string representation does
  not indicate multiple matrices and it's not immediately obvious
  what the RMF "plot" support will do

As a result of the backend changes, the error messages created
by the crates backend when handling "invalid" data (e.g. trying
to read a PHA as a RMF) may change (in this case to match those
that pyfits creates).

The testing is **very** limited, so there could well be problems
with the code.

At the moment we do not support writing out multi-matrix RMF
objects.