Address review comments by @hamogu
Mainly documentation but there's one code change: the class names
for the parameters has changed from

    Param and then IntParam based on it

to

    BaseParam and Param

since Param is going to be more-used than BaseParam and, as pointed
out in the review, IntParam is confusing and should be named
InterpolatedParameter, but that is just too ling, even if you have
auto-completion in your editor!

I didn't even toy with the idea of usong AddParam rather than
BaseParam, since they are identified as "additional" parameters,
because I didn't think it was a good name.