IO: move helper types into the types module and improvements
Move some of the helper types added in #1861 for handling XSPEC table
models out into the types module. This is not meant to be a generic
implementation of FITS-like capabilities, nor is it designed to be
efficient, it is just designed to be an interface layer between the
generic IO code and the backend-specific code.

As part of this, improve the capabilities of these types - e.g.
tracking min and max values of a column so that we can automatically
set the TLMIN/MAX values when writing it out (backend specific code
needed). This is used to improve the multi-matrix writing test but
will become useful in later commits.

These "new" types are then used in the I/O for sending data between
the front and back ends (mainly visible in the pack_hdus and set_hdus
routines, but some of the internal back-end code is updated to use
these "new" types).

There are a couple of other changes

- use importlib.metadata to access the Sherpa version number rather
  than use sherpa.__version__. There is some discussion that the
  former is the better way to do this - e.g.
  https://discuss.python.org/t/please-make-package-version-go-away/58501
  so try this new approach

- use contextlib.suppress to handle a "want to ignore an ImportError"
  situation; this is not any better (and goes against the "there's only
  one way to do it" ethos in Python) but I think it is a small
  improvement here (i.e. stating up front what errors we are happy to
  "ignore" rather than require the user to read to the end to find out,
  although it is only a minimal change).