IO: move helper types into the io_types module and improvements
Move some of the helper types added in #1861 for handling XSPEC table
models out into their own module. This is not meant to be a generic
implementation of FITS-like capabilities, nor is it designed to be
efficient, it is just designed to be an interface layer between the
generic IO code and the backend-specific code.

Allow the optional fields to not be given when constructing the
HeaderItem and Column types.

Support the FITS TLMIN/MAX standard for columns. This currently only
affects the test RMF we write out to test of the multi-rmf handling.

Improve typing of headers.

Move from the TableHDU type to Block/TableBlock/ImageBlock. It's too
early to see if it's a good idea and complicates the typing rules;
the BlockList type has been added to try and work around this.

Not all features are used yet.