IO: move RMF processing into the generic layer
Using io_types we can set up the data we want to both read in
and write out in the IO layer and not the backend layer.

One aim is to provide a unified interface for RMF processing
so that we no-longer have the case of "backend a can not read
in this RMF but backend b can" (modulo those cases where a
backend can not read in a particular file).

One minor change is that we now add a CREATOR keyword to the two
table blocks of a RMF (set to "sherpa <version>" or "sherpa"
if there's a problem accesing the version) and this is only
done if a CREATOR keyword is not already set.

There is some logic here that could be moved into the DataRMF
class as it looks like we can slow down the rmf_fold code
if the data types do not match (e.g. to make sure we convert
to the right data types needed by the C++ rmf_fold code).

RMF files are now required to follow the OGIP standard **much
more** strongly than before. For example, we now **require** an
EBOUNDS block when it was optional before. This might affect
some "interestig" use cases, but I honestly don't understand the
current support for this and it is completely untested (as shown
by the fact that making EBOUNDS a required block does not change
any existing test).

As part of this, we now read in all the MATRIX blocks for a RMF
and then warn the user we only use the first block (and this
warning is done in the generic layer rather than repeated in both
backends).

The typing is getting fun here.