TEST: improve coverage for RMF/PHA input
Check that we get a message when reading in the sis0.rmf file
about the lack of offset data (no TLMIN on F_CHAN or CHANNEL
columns).

Check that we get a warning about a missing background file
(BACKFILE keyword) for a PHA file (note that we have two places
where we report issues for missing *FILE keywords and BACKFILE
is handled differently to ANCR/RESPFILE which we already cover).