IO: move PHA processing into the generic layer
The backend layer now just reads in the data with the logic for
handling the data (e.g. is it type I vs type II, do we need to
convert from RATE to "counts") now in the generic layer. The
TableHDU process is also used for writing out the data. Thanks to
how the code was written this also changes the set_table_data to
use TableHDU for passing data from the generic layer to the backend
code.

There is some interesting behaviour - e.g. we don't fully handle
the "columns can also be given as a scalar value if it is a
constant" case, in particular with the GROUPING and QUALITY arrays.
We have at least one test file with them set to "0" and if we read
this in as a column then a number of tests fail. It is likely that
these failures can be treated as regression tests (i.e. the tests
can be updated) but that is not done here.

It is also a bit clearer (having done this) where the code now
exists that removes keywords like ANCRFILE from the header - see
issue #1885. However, it's still not entirely clear and the existing
code has some interesting behaviour (i.e. when writing out a PHA
the *FILE keywords have any paths stripped from them).

The pyfits backend had code that treated SYS_ERR as a fractional
error (i.e. it gets multiplied by the COUNTS column) but the crates
backend does not do this. So for now we drop this behaviour, even
though OGIP Memo OGIP/92-007 does say

  Sys_err, a 4-byte REAL scalar giving the fractional systematic
    error to be applied to the data in this row (channel).
  The FITS column name is SYS_ERR.

We do not have good testing of this column!

The code makes choices about columns and keywords it sets when
writing out PHA files. It's not entirely obvious all of these make
sense but the idea here is not to change the existing behaviour.
Any changes can come later.