IO: minor internal cleanup
User-visible change: remove a space at the end of a multi-line
warning message. This has long bothered my sensibilities so
remove it. We have tests that need to be updated because of this
change but it shows the tests are testing the thing we want testing.

Replace some old exception catching which used the sys module's
exception handling to diagnose the message to write out when we
can just grab it from the exception itself. We have tests that
catch this so we can check that - at least in the tested cases - it
works.

For PHA handlnig, move a global routine into _pack_pha as it is
only used there, and improving the comments.